Find/replace for VS2010 compatibility
On VS2010, template argument overloading isn't supported for types
that differ only in constness or pointer-type.  Therefore, the
templates in prelexer such as exactly<char>/exactly<const char *>
will cause VS2010 to error on compile.

This find/replace replaces the exactly<char> instantiation with a
renamed exactly_c<char>, which allows VS2010 to compile and test
successfully, albeit with a significant amount of warnings.